Unmap mapping in Select mode so it doesn’t interfere
`map` and `vmap`, surprisingly, apply not only to Visual mode but also to Select mode. (See `:help mapmode-s`.) However, this doesn’t make sense if the mapping contains a printable character, which it often does – it stops the user from typing that character and seeing the result immediately. So you have to use `sunmap` to remove it (or use `xmap` instead of `vmap`.)