Add support for detecting ancestors
Rooter starts with the directory of the buffer's file or directory and
works its way up the tree.  At each level it checks the patterns in
turn, looking for a match.

Previously the kinds of matches it could detect were is-a and has-a,
i.e. looking at the current level and its descendants.  Now it can also
look further up the tree and its ancestors.

See #101.